change how shade rows are capped
effectively removes the cap; same method as changing the max palette number.
anything above 8 shade slots will not be matched, accurate to how the game functions. it will also display a little [!] to the left of that shade row to indicate that the color data will not be shown.
this data, like the additional palette data past the maximum is actually stored by the game and usable with the functions get_profile_color_slot_r/g/b, which is something i personally use in my characters but have not seen used very often.